Fix typing
